# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Ennote ennoworker API
  version: 0.1.0
servers:
  - url: http://127.0.0.1:{port}
    variables:
      port:
        default: "0"
security:
  - bearerAuth: []
paths:
...
  /v1/roles:
    get:
      operationId: listRoles
      parameters:
        - { name: q, in: query, schema: { type: string, maxLength: 120 } }
        - { name: scope, in: query, schema: { $ref: "#/components/schemas/RoleScope" } }
        - { name: projectId, in: query, schema: { type: string } }
        - { name: flowId, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [active, archived], default: active } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
      responses:
        "200":
          description: Effective Role catalog without draft or prompt bodies
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties: { data: { $ref: "#/components/schemas/RoleCatalogPage" } }
    post:
      operationId: createRole
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRoleInput" }
      responses:
        "201":
          description: Role draft created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties: { data: { $ref: "#/components/schemas/RoleIdentity" } }
        "400": { $ref: "#/components/responses/Error" }
  /v1/roles/{roleID}:
    parameters:
      - { name: roleID, in: path, required: true, schema: { type: string } }
    get:
      operationId: getRole
      responses:
        "200":
          description: Role authoring detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties: { data: { $ref: "#/components/schemas/RoleIdentity" } }
        "404": { $ref: "#/components/responses/Error" }
...
  /v1/global-roles:
    get:
      operationId: listGlobalRoles
      summary: List Worker-global file-authored Roles
      responses:
        "200":
          description: Global Role source summaries
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties: { data: { type: array, items: { $ref: "#/components/schemas/GlobalRoleSummary" } } }
    post:
      operationId: createGlobalRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [document]
              properties: { document: { $ref: "#/components/schemas/PortableRoleDocument" } }
      responses:
        "201":
          description: Global Role source created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties: { data: { $ref: "#/components/schemas/GlobalRoleDetail" } }
        "422": { $ref: "#/components/responses/Error" }
  /v1/global-roles/{roleID}:
    parameters:
      - { name: roleID, in: path, required: true, schema: { type: string } }
    get:
      operationId: getGlobalRole
      responses:
        "200":
          description: Global Role source
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties: { data: { $ref: "#/components/schemas/GlobalRoleDetail" } }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      operationId: updateGlobalRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [expectedDigest, document]
              properties:
                expectedDigest: { type: string, pattern: "^sha256:[0-9a-f]{64}$" }
                document: { $ref: "#/components/schemas/PortableRoleDocument" }
      responses:
        "200":
          description: Updated global Role source
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties: { data: { $ref: "#/components/schemas/GlobalRoleDetail" } }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
...
  /v1/global-roles/{roleID}/publish:
    parameters:
      - { name: roleID, in: path, required: true, schema: { type: string } }
    post:
      operationId: publishGlobalRole
      responses:
        "201":
          description: Global Role source published as an immutable runtime version
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
        "422": { $ref: "#/components/responses/Error" }
  /v1/global-roles/{roleID}/versions:
    parameters:
      - { name: roleID, in: path, required: true, schema: { type: string } }
    get:
      operationId: listGlobalRoleRevisions
      responses:
        "200":
          description: Immutable file-backed Role revisions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/ImmutableFileRevision" }
        "404": { $ref: "#/components/responses/Error" }
  /v1/global-roles/{roleID}/resolved:
    parameters:
      - { name: roleID, in: path, required: true, schema: { type: string } }
    get:
      operationId: getResolvedGlobalRole
      summary: Resolve a bare Role handle to its latest published revision with audit source
      responses:
        "200":
          description: Resolved effective Role
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties: { data: { $ref: "#/components/schemas/ResolvedGlobalRole" } }
        "404": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    Envelope:
      type: object
      properties:
        data: {}
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId, retryable]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            retryable: { type: boolean }
...
    ThinkingEffort:
      type: string
      enum: [default, low, medium, high]
...
    ImmutableFileRevision:
      type: object
      additionalProperties: false
      required: [schemaVersion, resourceId, version, digest, publishedAt]
      properties:
        schemaVersion: { type: integer, enum: [1] }
        resourceId: { type: string }
        version: { type: integer, minimum: 1 }
        digest: { type: string, pattern: "^sha256:[0-9a-f]{64}$" }
        publishedAt: { type: string, format: date-time }
    GlobalRoleSummary:
      type: object
      additionalProperties: false
      required: [id, name, path]
      properties:
        id: { type: string }
        name: { type: string }
        path: { type: string }
        digest: { type: string }
        error: { type: string }
    PortableRoleDocument:
      type: object
      additionalProperties: false
      required: [schemaVersion, handle, name, description, positioning, icon, color, model, skills, authority, permissionCeiling, allowedTools, context, delegation, outputContract, maxLoopIterations, prompt]
      properties:
        schemaVersion: { type: integer, enum: [1] }
        handle: { type: string, pattern: "^[a-z][a-z0-9_-]{1,31}$" }
        name: { type: string, minLength: 1 }
        description: { type: string }
        positioning: { type: string }
        icon: { type: string }
        color: { type: string }
        model:
          type: object
          additionalProperties: false
          required: [ref, thinkingEffort, fallbacks]
          properties:
            ref: { type: string, pattern: "^[^/]+/[^/]+$" }
            thinkingEffort: { $ref: "#/components/schemas/ThinkingEffort" }
            fallbacks: { type: array, uniqueItems: true, items: { type: string } }
        skills:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [id, mode]
            properties:
              id: { type: string }
              mode: { type: string, enum: [preload, available] }
        authority: { type: string, enum: [read_only, write_workspace] }
        permissionCeiling: { type: string, enum: [discuss, ask, auto] }
        allowedTools: { type: array, uniqueItems: true, items: { type: string } }
        context:
          type: object
          additionalProperties: false
          required: [defaultMode, allowedModes, ownExecutionContinuity]
          properties:
            defaultMode: { type: string, enum: [room, fresh] }
            allowedModes: { type: array, uniqueItems: true, items: { type: string, enum: [room, fresh] } }
            ownExecutionContinuity: { type: string, enum: [none] }
        delegation:
          type: object
          additionalProperties: false
          required: [admission, allowedCallerKinds, allowedStrategies, maxInvocationsPerParentRun, maxConcurrentInstances, budgetCeiling]
          properties:
            admission: { type: string, enum: [auto_within_budget, approval_required, deny] }
            allowedCallerKinds: { type: array, uniqueItems: true, items: { type: string } }
            allowedStrategies: { type: array, uniqueItems: true, items: { type: string } }
            maxInvocationsPerParentRun: { type: integer, minimum: 1 }
            maxConcurrentInstances: { type: integer, minimum: 1 }
            budgetCeiling:
              type: object
              additionalProperties: false
              required: [maxModelCalls, maxToolCalls, maxTotalTokens, maxOutputTokens, maxCostUsdMicros, maxWallTimeMs]
              properties:
                maxModelCalls: { type: integer, minimum: 1 }
                maxToolCalls: { type: integer, minimum: 0 }
                maxTotalTokens: { type: integer, minimum: 1 }
                maxOutputTokens: { type: integer, minimum: 1 }
                maxCostUsdMicros: { type: integer, minimum: 0 }
                maxWallTimeMs: { type: integer, minimum: 1 }
        outputContract: { type: string }
        maxLoopIterations: { type: integer, minimum: 1 }
        prompt: { type: string, minLength: 1, maxLength: 65536 }
    GlobalRoleDetail:
      allOf:
        - $ref: "#/components/schemas/GlobalRoleSummary"
        - type: object
          required: [digest, document]
          properties:
            digest: { type: string }
            document: { $ref: "#/components/schemas/PortableRoleDocument" }
    ResolvedGlobalRole:
      type: object
      additionalProperties: false
      required: [roleId, source, revision, document]
      properties:
        roleId: { type: string }
        source: { type: string }
        revision: { $ref: "#/components/schemas/ImmutableFileRevision" }
        document: { $ref: "#/components/schemas/PortableRoleDocument" }
...
    RoleScope:
      type: string
      enum: [builtin, global, project, flow]
    RoleModelBinding:
      type: object
      additionalProperties: false
      required: [mode, thinkingEffort, fallbackModelProfileIds, overridableFields]
      properties:
        mode: { type: string, enum: [fixed, overridable, inherit] }
        modelProfileId: { type: string }
        thinkingEffort: { $ref: "#/components/schemas/ThinkingEffort" }
        fallbackModelProfileIds: { type: array, uniqueItems: true, items: { type: string } }
        overridableFields: { type: array, uniqueItems: true, items: { type: string } }
    RoleSkillEntry:
      type: object
      additionalProperties: false
      required: [skillId, mode]
      properties:
        skillId: { type: string }
        mode: { type: string, enum: [preload, available] }
...
    DelegationBudgetCeiling:
      type: object
      additionalProperties: false
      required: [maxModelCalls, maxToolCalls, maxTotalTokens, maxOutputTokens, maxCostUsdMicros, maxWallTimeMs]
      properties:
        maxModelCalls: { type: integer, minimum: 0 }
        maxToolCalls: { type: integer, minimum: 0 }
        maxTotalTokens: { type: integer, format: int64, minimum: 0 }
        maxOutputTokens: { type: integer, format: int64, minimum: 0 }
        maxCostUsdMicros: { type: integer, format: int64, minimum: 0 }
        maxWallTimeMs: { type: integer, format: int64, minimum: 0 }
    RoleDelegationPolicy:
      type: object
      additionalProperties: false
      required: [admission, allowedCallerKinds, allowedStrategies, maxInvocationsPerParentRun, maxConcurrentInstances, budgetCeiling]
      properties:
        admission: { type: string, enum: [denied, approval_required, auto_within_budget] }
        allowedCallerKinds:
          type: array
          uniqueItems: true
          items: { type: string, enum: [host] }
        allowedStrategies:
          type: array
          uniqueItems: true
          items: { type: string, enum: [single, parallel] }
        maxInvocationsPerParentRun: { type: integer, minimum: 0 }
        maxConcurrentInstances: { type: integer, minimum: 0 }
        budgetCeiling: { $ref: "#/components/schemas/DelegationBudgetCeiling" }
    RoleContextPolicy:
      type: object
      additionalProperties: false
      required: [defaultMode, allowedModes, ownExecutionContinuity]
      properties:
        defaultMode: { type: string, enum: [room, reply, fresh, task_only] }
        allowedModes:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: string, enum: [room, reply, fresh, task_only] }
        ownExecutionContinuity: { type: string, enum: [none] }
    RoleDefinition:
      type: object
      additionalProperties: false
      required: [schemaVersion, rolePrompt, modelBinding, skills, authority, permissionCeiling, allowedTools,
        contextPolicy, delegationPolicy, outputContract, maxLoopIterations]
      properties:
        schemaVersion: { type: integer, enum: [1] }
        rolePrompt: { type: string, minLength: 1, maxLength: 65536 }
        modelBinding: { $ref: "#/components/schemas/RoleModelBinding" }
        skills:
          type: object
          additionalProperties: false
          required: [entries]
          properties:
            entries: { type: array, items: { $ref: "#/components/schemas/RoleSkillEntry" } }
        authority: { type: string, enum: [read_only, mutation] }
        permissionCeiling: { type: string, enum: [discuss, ask, auto] }
        allowedTools: { type: array, uniqueItems: true, items: { type: string } }
        contextPolicy: { $ref: "#/components/schemas/RoleContextPolicy" }
        delegationPolicy: { $ref: "#/components/schemas/RoleDelegationPolicy" }
        outputContract: { type: string, minLength: 1 }
        maxLoopIterations: { type: integer, minimum: 1, maximum: 64 }
    RoleSummary:
      type: object
      additionalProperties: false
      required: [id, handle, name, description, positioning, icon, color, scope, status, sourceKind, updatedAt]
      properties:
        id: { type: string }
        handle: { type: string }
        name: { type: string }
        description: { type: string }
        positioning: { type: string }
        icon: { type: string }
        color: { type: string }
        scope: { $ref: "#/components/schemas/RoleScope" }
        projectId: { type: string }
        flowId: { type: string }
        status: { type: string, enum: [active, archived] }
        sourceKind: { type: string, enum: [managed, project_file] }
        sourceLocator: { type: string, description: Informational Worker-resolved source path; never accepted from clients }
        currentVersionId: { type: string }
        currentVersion: { type: integer, minimum: 1 }
        updatedAt: { type: string, format: date-time }
    RoleIdentity:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/RoleSummary"
        - type: object
          required: [draft, draftRevision, delegationEnabled, delegationRevocationEpoch, createdAt]
          properties:
            sourceDigest: { type: string, pattern: "^sha256:[0-9a-f]{64}$" }
            draft: { $ref: "#/components/schemas/RoleDefinition" }
            draftRevision: { type: integer, minimum: 0 }
            delegationEnabled: { type: boolean }
            delegationRevocationEpoch: { type: integer, minimum: 0 }
            delegationDisabledAt: { type: string, format: date-time }
            createdAt: { type: string, format: date-time }
...
    RoleCatalogPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/RoleSummary" } }
        nextCursor: { type: string }
...
    CreateRoleInput:
      type: object
      additionalProperties: false
      required: [handle, name, scope, definition]
      properties:
        handle: { type: string, pattern: "^[a-z][a-z0-9_-]{1,31}$" }
        name: { type: string, minLength: 1 }
        description: { type: string }
        positioning: { type: string }
        icon: { type: string }
        color: { type: string }
        scope: { $ref: "#/components/schemas/RoleScope" }
        projectId: { type: string }
        flowId: { type: string }
        definition: { $ref: "#/components/schemas/RoleDefinition" }
...